# auth-service/swagger.yaml
openapi: 3.0.0
info:
  title: API d'Authentification
  version: 1.0.0
  description: API pour la gestion de l'authentification des utilisateurs

servers:
  - url: http://localhost:3001
    description: Serveur de développement

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        isEmailVerified:
          type: boolean
        isPhoneVerified:
          type: boolean
        twoFactorEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

paths:
  /auth/register:
    post:
      summary: Inscription d'un nouvel utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                password:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                phoneNumber:
                  type: string
      responses:
        '201':
          description: Utilisateur créé avec succès
        '400':
          description: Données invalides

  /auth/login:
    post:
      summary: Connexion d'un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Connexion réussie
        '401':
          description: Identifiants invalides

  /auth/google:
    get:
      summary: Authentification avec Google
      responses:
        '302':
          description: Redirection vers Google

  /auth/google/callback:
    get:
      summary: Callback Google OAuth
      responses:
        '302':
          description: Redirection après authentification

  /users/me:
    get:
      summary: Obtenir le profil de l'utilisateur
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Non authentifié

    put:
      summary: Mettre à jour le profil
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Profil mis à jour
        '401':
          description: Non authentifié

    delete:
      summary: Supprimer le compte
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Compte supprimé
        '401':
          description: Non authentifié

  /users/2fa:
    put:
      summary: Activer/désactiver la double authentification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - enabled
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Statut 2FA mis à jour
        '401':
          description: Non authentifié 